requirements.txt
bleak==2.1.1
dbus-fast==4.0.0
python-dotenv==1.2.2
